Introduce FaLayersTextDirective and refactor FaLayersTextComponent to use host directives
